Add license header for Jrag.jjt
The Jrag.jjt file lacked license header. Jrag.jjt is based on the Java1.1.jjt
file from the VTransformer example in JavaCC. The Jrag.jjt file is based on an
old release of JavaCC from 1997, but the same file was relicensed to Modified
BSD License by Sun Microsystems in JavaCC 5.0. We include the Modified BSD
License header and copyright notice from the JavaCC 5.0 release. A comment is
left to document the history of the Jrag.jjt file.