Remove unused "refine to refine" syntax
This change simplifies both the refinement processing and parsing of refine
statements.

The refinement syntax for methods and synthesized equations had a variation
where the equation or code block was replaced by "to refine", for example:

refine Aspect1 eq A.x() to refine Aspect2 eq A.y() = ...;

This could be used to rename an attribute, however a similar effect can be
achieved by simply creating a new attribute calling the old one. The only
difference is that the old attribute remains.