Replacement of Empty Container Nodes by Singleton
The main idea is to check in setChild that all Opt and List nodes which
have no children get pointed to the singletons List.EMPTY and Opt.EMPTY.
The singletons have overridden versions of mutating methods that throw
errors when they are called. Additionally, the mutator methods in the
parent of a List node now include a special case, where it is checked
whether the current node is the singleton node before appending to it.