Fix NTA initialization recursion in concurrent mode
NTA proxy object initialization would crash in recursion in concurrent mode for
parameterized NTAs declared on ASTNode because the proxy object initialization
was not lazy.

This changes NTA proxy object initialization to be lazy, as in the
non-concurrent mode. This is done with AtomicReference to avoid initialization
data race.

fixes #287 (bitbucket)