Marks incremental/region handlers as computed when added as listeners.
Handlers are added as listeners when regions are crossed, that is, when parents
and children at the region edge are accessed. After such a crossing values in the
entered region may be used freely, hence the region accessing the neighbouring
region is added as a listener and is also marked as computed.