Adds creation of child handler in the List constructor for incremental/param.
For cases when a list is created via the List(T... ) constructor. This constructor
creates a child array and then calls addChild (which calls setChild). In setChild,
the creation of the child handler is controlled by whether the child array is null.
Creating the child handler in this constructor ensures that it is no null.