Adds change notification to collection attributes in the same group for incremental.
Given that collection attributes are computed in batch, the first collection attribute
instance computed for a group will collect all the dependencies for the group of
collection attributes. Hence, at time of flush, the whole group of collection attributes
must be flushed and not only the notified initial collection attribute instance driving
the evaluation for the group.