Fix error in NTA child storage for concurrent mode
When there was not an equation for some NTA on an AST class, but the NTA hand
an equation in some superclass, the NTA was incorrectly accessed through the
child vector instead of having a special accessor through getChildNoTransform().
This error was just caused by not looking for NTA equations in superclasses
in ASTDecl.hasSynEq().

fixes #299 (bitbucket)